Add basic templating and configuration
Add a basic configuration interface so that the text and the URL
aren't hardcoded.

Also, add basic templating to the text and URL values. Similar to that of
the table plugin, but only supporting "${__cell_XXX}", where XXX is a
number.

While at it, make it look like it should (big font and all that).